Object creation logic w/ switch case
Trying to control the creation of the "drugs" and "peaces" with switch case. The idea is to have a fixed logic in the beginning and as long as the player fails or succeeds by getting "drugs" or "peaces" the logic modifies increasing the speed or the amount of "drugs" created. Also decrease and fasten the "peaces"